Modularize and relocate all non-core Whisper driver components into a new `support` module to streamline and segregate functionality boundaries. Update all associated imports, documentation, and tests accordingly.
