+++ crates/public/bunsen/src/kits/speech/whisper/decode/mod.rs
@@ -41,7 +41,7 @@ use crate::{
             Whisper,
             WhisperMeta,
         },
-        tokens::WhisperSpecialIds,
+        driver::support::WhisperSpecialIds,
     },
     ops::{
         repeat::repeat_interleave,
