+++ crates/public/bunsen/src/kits/speech/whisper/driver/support/emission.rs
