+++ crates/public/bunsen/src/kits/speech/whisper/driver/support/mod.rs
@@ -1,0 +1,33 @@
+//! Support functions for the Whisper driver.
+
+mod clamp;
+mod clock;
+mod emission;
+mod mel;
+mod regions;
+mod segments;
+mod text;
+mod tokens;
+mod va_filter;
+mod vocab;
+
+#[doc(inline)]
+pub use clamp::*;
+#[doc(inline)]
+pub use clock::*;
+#[doc(inline)]
+pub use emission::*;
+#[doc(inline)]
+pub use mel::*;
+#[doc(inline)]
+pub use regions::*;
+#[doc(inline)]
+pub use segments::*;
+#[doc(inline)]
+pub use text::*;
+#[doc(inline)]
+pub use tokens::*;
+#[doc(inline)]
+pub use va_filter::*;
+#[doc(inline)]
+pub use vocab::*;
