+++ crates/validation/whisper-model-validation/src/audio/download/gpu_tests.rs
@@ -3,11 +3,13 @@ use bunsen::{
         ApplyTimestampRules,
         DecodeConfig,
         GreedyDecodeConfig,
-        MaxSeen,
-        Task,
-        TimestampHistory,
         WhisperDriverConfig,
         default_filters,
+        driver::support::{
+            MaxSeen,
+            Task,
+            TimestampHistory,
+        },
         mel_windows,
     },
     prelude::TensorElemOpExt,
