+++ crates/validation/whisper-model-validation/src/audio/mod.rs
@@ -35,14 +35,16 @@ use bunsen::{
     },
     kits::{
         speech::whisper::{
-            Task,
-            TiktokenRanks,
-            TokenPolicy,
             Whisper,
-            mel_options,
-            package_mels,
+            driver::support::{
+                Task,
+                TiktokenRanks,
+                TokenPolicy,
+                detokenizer,
+                mel_options,
+                package_mels,
+            },
             pretrained::bundled,
-            text::detokenizer,
         },
         tokens::{
             Detokenizer,
