+++ examples/whisper-cli/src/main.rs
@@ -17,20 +17,22 @@ use bunsen::{
     kits::speech::{
         silero_vad::SileroVad,
         whisper::{
-            Emission,
-            EmissionPolicy,
             FallbackConfig,
-            MaxSeen,
             SAMPLE_RATE,
-            Task,
-            TimestampHistory,
-            TokenPolicy,
-            VoiceActivityFilterConfig,
             Whisper,
             WhisperDriver,
             WhisperDriverConfig,
             decode::default_filters,
-            detokenizer,
+            driver::support::{
+                Emission,
+                EmissionPolicy,
+                MaxSeen,
+                Task,
+                TimestampHistory,
+                TokenPolicy,
+                VoiceActivityFilterConfig,
+                detokenizer,
+            },
             pretrained::bundled_vocabulary,
         },
     },
